requirements_slurm.txt
# Core training/runtime deps (installed into ./venv by submit_*.sh)
# NOTE: torch/torchvision/torchaudio are installed separately in the SLURM script
# to ensure CUDA-enabled wheels.

diffusers>=0.36.0
transformers>=4.46.3
accelerate>=0.30.0
safetensors>=0.4.0
huggingface-hub>=0.23.0

omegaconf>=2.3.0
pyyaml>=6.0
python-dotenv>=1.0.0

numpy
Pillow
tqdm

# Tokenizer dependencies for SD3
sentencepiece>=0.1.99
protobuf>=3.20.0
